backend/requirements.txt
# FastAPI
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.10.0
psycopg2-binary>=2.9.5

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Firebase
firebase-admin>=6.1.0

# HTTP Client
httpx>=0.23.0

# Data processing and prediction models
numpy>=1.24.2
pandas>=2.0.0
scikit-learn>=1.2.2
statsmodels>=0.14.0

# Logging
python-json-logger>=2.0.7

# Email
emails>=0.6

# Timezone
pytz>=2023.0